requirements.txt
asgiref==3.6.0
colorama==0.4.6
django-autoslug==1.9.8
django-cities-light==3.9.2
django-phonenumber-field==7.0.2
Django==4.1.5
loguru==0.6.0
phonenumbers==8.13.4
Pillow==9.4.0
pip==21.3.1
progressbar2==4.2.0
python-utils==3.4.5
pytz==2022.7.1
setuptools==60.2.0
sqlparse==0.4.3
tzdata==2022.7
Unidecode==1.3.6
wheel==0.37.1
win32-setctime==1.1.0
